adding decorators to flask views for user calls. Added error messages to its own class and still adding them into server error messages. Fixed a bug with edit address not setting the same address to new default if original was

+++ api/models/user.py
@@ -277,8 +277,6 @@ class User(db.Model):
 		try:
 			addresses = Lob.getUserAddresses(self)
 			sorted_addresses = sorted(addresses,  key=lambda k: k['date_created'])
-			# for address in sorted_addresses:
-			# 	print(address['address_zip'])
 			return sorted_addresses
 		except:
 			return []
@@ -287,11 +285,19 @@ class User(db.Model):
 	def editAddress(self, address_id, description, name, address_line1, address_line2, address_city, address_state,
 			address_zip, address_country):
 
+		old_address = Lob.getAddressById(address_id)
+		if int(old_address.metadata.get(Labels.AccountId)) != self.account_id:
+			raise Exception("This address belongs to a different user")
 		Lob.deleteAddress(address_id)
 		address = Lob.addUserAddress(self, description = description, name = name, address_line1 = address_line1
 			, address_line2 = address_line2, address_city = address_city,
 				address_state = address_state, address_zip = address_zip,
 				 address_country = address_country)
+
+		if self.default_address == address_id:
+			self.default_address = address['id']
+			db.session.commit()
+
 		return address
